part 1 (BACKEND)/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.5
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0